Added SSO
Added SSO

+++ database/migrations/2022_01_19_193207_create_reviews_table.php
@@ -18,7 +18,7 @@ class CreateReviewsTable extends Migration
             $table->timestamps();
             $table->foreignId('beerId')->constrained('beers');
             $table->text('review');
-            $table->integer('score');
+            $table->float('score');
             $table->foreignId('userId')->constrained('users');
         });
     }
